+++ src/index.ts
@@ -1,51 +1,4 @@
-import js from '@eslint/js'
-import ts from '@typescript-eslint/eslint-plugin'
-import tsParser from '@typescript-eslint/parser'
 import type { Linter } from 'eslint'
-import prettier from 'eslint-config-prettier'
-import deprecation from 'eslint-plugin-deprecation'
 
-export const tsConfig = (options?: {
-  exclude?: {
-    deprecation?: boolean
-    prettier?: boolean
-  }
-  tsConfigPath?: string
-  ignores?: string[]
-}): Linter.FlatConfig[] => [
-  {
-    files: ['**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}']
-  },
-  {
-    ignores: options?.ignores ?? ['dist', '*.config.{js,ts,cjs,mjs,cts,mts}']
-  },
-  {
-    plugins: {
-      // @ts-expect-error workaround until upstream update
-      '@typescript-eslint': ts
-    },
-    languageOptions: {
-      // @ts-expect-error workaround until upstream update
-      parser: tsParser,
-      parserOptions: {
-        project: [options?.tsConfigPath ?? './tsconfig.json']
-      }
-    },
-    // @ts-expect-error workaround until upstream update
-    rules: {
-      ...ts.configs.base.rules,
-      ...ts.configs.recommended.rules
-    }
-  },
-  js.configs.recommended,
-  options?.exclude?.prettier ? {} : prettier,
-  // @ts-expect-error workaround until upstream update
-  options?.exclude?.deprecation
-    ? {}
-    : {
-        plugins: {
-          deprecation
-        },
-        rules: deprecation.configs.recommended.rules
-      }
-]
+export { tsConfig } from './tsConfig.js'
+export type FlatConfig = Linter.FlatConfig
